KA-468 Rename keys.js file
