+++ kcd-search-service-update/index.js
@@ -1,6 +1,6 @@
 const getCodenamesOfItems = require('../shared/utils/codenamesExtractor');
 const indexers = require('../shared/searchIndexers');
-const appKeys = require('../shared/external/keys');
+const { setupConfiguration } = require('../shared/external/configuration');
 
 function validateEvent(event) {
     return event.eventType === 'kentico' &&
@@ -10,7 +10,7 @@ function validateEvent(event) {
 
 module.exports = async (context, eventGridEvent) => {
     if (validateEvent(eventGridEvent)) {
-        appKeys.setupConfiguration(eventGridEvent.isTest);
+        setupConfiguration(eventGridEvent.isTest);
         const codenames = getCodenamesOfItems(eventGridEvent.data.items, 'article');
         await indexers.reindexSpecificArticles(codenames);
     } else {
