+++ shared/external/configuration.js
